requirements.txt
google-genai
dotenv
matplotlib
pandas
scikit-learn
seaborn
plotly